Move request logic to worker
