+++ bin/worker.js
@@ -1,6 +1,13 @@
 import { handleRequest } from "../lib/request";
 import { reportMetric } from "../lib/metrics";
 
+export const handleRequest = async (request, env) => {
+  const url = new URL(request.url);
+  const newRequest = new Request(`${env.REQUEST_URL}${url.pathname}${url.search}`, request);
+
+  return await fetch(newRequest);
+}
+
 export default {
   async fetch(request, env, ctx) {
     let response;
