+++ test/request.test.js
@@ -1,5 +1,5 @@
-import { describe, it, expect, vi } from 'vitest';
-import { formatRequestData, handleRequest } from '../lib/request.js';
+import { describe, it, expect } from 'vitest';
+import { formatRequestData } from '../lib/request.js';
 
 describe('formatRequestData', () => {
   it('should format request data correctly', () => {
@@ -30,31 +30,5 @@ describe('formatRequestData', () => {
     expect(result).toContain('method=GET');
     expect(result).toContain('cf_cache=HIT');
     expect(result).toContain('api_key=test-key');
-  });
-});
-
-
-
-describe('handleRequest', () => {
-  it('should handle request correctly', async () => {
-    const request = {
-      url: 'https://example.com/path?api-key=test-key',
-      method: 'GET'
-    };
-    const env = {
-      REQUEST_URL: 'https://proxy.example.com'
-    };
-
-    global.fetch = vi.fn().mockResolvedValue({ status: 200 });
-
-    const response = await handleRequest(request, env);
-
-    expect(global.fetch).toHaveBeenCalledWith(
-      expect.objectContaining({
-        url: 'https://proxy.example.com/path?api-key=test-key',
-        method: 'GET'
-      })
-    );
-    expect(response.status).toBe(200);
   });
 });
